% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-post_strat_estimator.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate post-stratification weights}
\usage{
calculate_weights(pop_df, u5, p)
}
\description{
Calculate post-stratification weights
}
\keyword{internal}
