% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_coverage_data}
\alias{check_coverage_data}
\title{Check coverage data for post-stratification weighted estimation}
\usage{
check_coverage_data(cov_df)
}
\arguments{
\item{cov_df}{A \code{\link[=data.frame]{data.frame()}} of stratified coverage survey data.}
}
\value{
A message or an error on whether \code{cov_df} is structured
appropriately for post-stratification weighted estimation.
}
\description{
Check coverage data for post-stratification weighted estimation
}
\keyword{internal}
