% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pop_data}
\alias{check_pop_data}
\title{Check population data for post-stratification weighted estimation}
\usage{
check_pop_data(pop_df)
}
\arguments{
\item{pop_df}{A \code{\link[=data.frame]{data.frame()}} of population data.}
}
\value{
A message or an error on whether \code{pop_df} is structured
appropriately for post-stratification weighted estimation.
}
\description{
Check population data for post-stratification weighted estimation
}
\keyword{internal}
