% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-classify_coverage.R
\name{classify_coverage}
\alias{classify_coverage}
\title{Classify coverage results}
\usage{
classify_coverage(n_in, n_total, standard = c(0.2, 0.5))
}
\arguments{
\item{n_in}{Number (integer) of cases found in the programme}

\item{n_total}{Number (integer) of children under 5 years sampled}

\item{standard}{Decision rule standard/s. Should be between 0 and 1. At
least one standard should be provided for a two-tier coverage classifier.
Two standards should be provided for a three-tier coverage classifier.
Default is a three-tier classifier with rule set at 0.2 and 0.5.}
}
\value{
A character value or vector indicating coverage classification. If
\code{standard} is a single value, returns \strong{"Satisfactory"} if coverage is
above \code{standard} and \strong{"Not satisfactory"} if coverage is below or
equal to \code{standard}. If \code{standard} is two values, returns \strong{"Low"} if
coverage is below or equal to lower standard, \strong{"High"} if coverage is
above the higher standard, and \strong{"Moderate"} for all other coverage
values.
}
\description{
Classify coverage results
}
\examples{
classify_coverage(n_in = 6, n_total = 40)
with(survey_data,
  classify_coverage(n_in = in_cases, n_total = n)
)

}
\author{
Ernest Guevarra
}
