% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_binom_hypergeom}
\alias{get_binom_hypergeom}
\title{Calculate the binomial coefficient \emph{n-choose-k}}
\usage{
get_binom_hypergeom(n, k)
}
\arguments{
\item{n}{Total population}

\item{k}{Number of sample drawn from total population}
}
\value{
A numeric value of binomial probability
}
\description{
Calculate the binomial coefficient \emph{n-choose-k}
}
\keyword{internal}
