% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-test_classifier.R
\name{get_class_prob}
\alias{get_class_prob}
\title{Function to produce misclassification probabilities}
\usage{
get_class_prob(x)
}
\arguments{
\item{x}{Simulated results data produced by \code{\link[=test_lqas_classifier]{test_lqas_classifier()}}}
}
\value{
A list of LQAS misclassification probabilities results
}
\description{
Function to produce misclassification probabilities
}
\examples{
sim <- test_lqas_classifier(replicates = 5, runs = 5,
                            pop = 10000, n = 40,
                            d.lower = 60, d.upper = 90)
get_class_prob(x = sim)

}
