% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_classification_probabilities}
\alias{get_classification_probabilities}
\title{Get low, moderate, and high probabilities per proportion}
\usage{
get_classification_probabilities(x)
}
\arguments{
\item{x}{A vector of low, moderate, and high classification labelled as
1, 2, and 3 respectively.}
}
\value{
A table object of low, moderate, and high classification
probabilities
}
\description{
Get low, moderate, and high probabilities per proportion
}
\keyword{internal}
