% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-sample_size.R
\name{get_d}
\alias{get_d}
\title{Calculate decision rule for a specified sample size and lower and upper
triage thresholds}
\usage{
get_d(N, n, dLower, dUpper, alpha = 0.1, beta = 0.1)
}
\arguments{
\item{N}{Total population size of cases in the specified survey area}

\item{n}{Sample size}

\item{dLower}{Lower triage threshold. Values from 0 to 1.}

\item{dUpper}{Upper triage threshold. Values from 0 to 1.}

\item{alpha}{Maximum tolerable alpha error. Values from 0 to 1.
Default is 0.1}

\item{beta}{Maximum tolerable beta error. Values from 0 to 1. Default is 0.1}
}
\value{
A list of values providing the LQAS sampling plan for the specified
  parameters. The list includes sample size, decision rule, alpha error and
  beta error for the specified classification scheme
}
\description{
Calculate decision rule for a specified sample size and lower and upper
triage thresholds
}
\examples{
get_d(N = 600, n = 40, dLower = 0.7, dUpper = 0.9)

}
