% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-sample_size.R
\name{get_hypergeom_cumulative}
\alias{get_hypergeom_cumulative}
\title{Calculate cumulative hypergeometric probabilities}
\usage{
get_hypergeom_cumulative(k, m, n, N, tail = "lower")
}
\arguments{
\item{k}{Number of cases in the sample}

\item{m}{Number of cases in the population}

\item{n}{Sample size}

\item{N}{Population size}

\item{tail}{A character vector indicating "lower" (default) or "upper" tail}
}
\description{
Calculate cumulative hypergeometric probabilities
}
\examples{
get_hypergeom_cumulative(k = 5, m = 600, n = 25, N = 10000)

}
