% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-sample_size.R
\name{get_n}
\alias{get_n}
\title{Calculate sample size of number of cases to be found to assess coverage}
\usage{
get_n(c, dLower, dUpper, alpha = 0.1, beta = 0.1)
}
\arguments{
\item{c}{Total population size of cases in the specified survey area}

\item{dLower}{Lower triage threshold. Values from 0 to 1.}

\item{dUpper}{Upper triage threshold. Values from 0 to 1.}

\item{alpha}{Maximum tolerable alpha error. Values from 0 to 1.
Default is 0.1}

\item{beta}{Maximum tolerable beta error. Values from 0 to 1. Default is 0.1}
}
\value{
A list of values providing the LQAS sampling plan for the specified
  parameters. The list includes sample size, decision rule, alpha error and
  beta error for the specified classification scheme
}
\description{
Calculate sample size of number of cases to be found to assess coverage
}
\examples{
get_n(c = 600, dLower = 0.7, dUpper = 0.9)

}
