% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-test_classifier.R
\name{lqas_get_class_prob}
\alias{lqas_get_class_prob}
\title{Produce misclassification probabilities}
\usage{
lqas_get_class_prob(x)
}
\arguments{
\item{x}{Simulated results data produced by \code{\link[=lqas_simulate_test]{lqas_simulate_test()}}}
}
\value{
A list of LQAS misclassification probabilities results
}
\description{
Produce misclassification probabilities
}
\examples{
sim <- lqas_simulate_test(
  pop = 10000, n = 40, dLower = 0.6, dUpper = 0.9, replicates = 5, runs = 5
)

lqas_get_class_prob(x = sim)

}
