% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-test_classifier]].R
\name{lqas_run}
\alias{lqas_run}
\title{Function to perform LQAS based on data based on specified decision rules}
\usage{
lqas_run(data, n, d.lower, d.upper)
}
\arguments{
\item{data}{A vector of simulated data produced by \code{make_data}}

\item{n}{Sample size of actual or test coverage data}

\item{d.lower}{A numeric value for the lower classification threshold}

\item{d.upper}{A numeric value for the upper classification threshold}
}
\value{
A list of coverage proportions and LQAS outcomes
}
\description{
Function to perform LQAS based on data based on specified decision rules
}
\examples{
lqas_run(data = make_data(proportion = 0.3, pop = 10000),
         n = 40, d.lower = 60, d.upper = 90)

}
