% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-test_classifier.R
\name{make_data}
\alias{make_data}
\title{Function to simulate survey data of covered/cases and non-covered/non-cases
given specified parameters}
\usage{
make_data(proportion, pop)
}
\arguments{
\item{proportion}{A numeric value of a coverage proportion to simulate on}

\item{pop}{Population size from which simulated coverage survey data is to
be taken from}
}
\value{
A numeric vector of cases and non-cases (1s and 0s)
}
\description{
Function to simulate survey data of covered/cases and non-covered/non-cases
given specified parameters
}
\examples{
make_data(proportion = 0.3, pop = 10000)

}
