% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{Population data of districts of Sierra Leone based on 2015 census}
\format{
A tibble with 14 rows and 2 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{district} \tab District name \cr
   \emph{pop} \tab Population \cr
}
}
\source{
https://sierraleone.unfpa.org/sites/default/files/pub-pdf/Population\%20structure\%20Report_1.pdf
}
\usage{
pop_data
}
\description{
Population data of districts of Sierra Leone based on 2015 census
}
\examples{
pop_data

}
\keyword{datasets}
