% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-test_classifier.R
\name{print.lqasClass}
\alias{print.lqasClass}
\title{\code{print} helper function for \code{get_class_prob} function}
\usage{
\method{print}{lqasClass}(x, ...)
}
\arguments{
\item{x}{An object resulting from applying the \code{get_class_prob()}
function.}

\item{...}{Additional \code{print()} parameters}
}
\value{
Printed output of \code{get_class_prob()} function
}
\description{
\code{print} helper function for \code{get_class_prob} function
}
\examples{
sim <- test_lqas_classifier(replicates = 5, runs = 5,
                            pop = 10000, n = 40,
                            d.lower = 60, d.upper = 90)
x <- get_class_prob(x = sim)
print(x)

}
