% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-test_classifier.R
\name{run_lqas}
\alias{run_lqas}
\title{Function to perform LQAS based on data based on specified decision rules}
\usage{
run_lqas(data, n, d.lower, d.upper)
}
\arguments{
\item{data}{A vector of simulated data produced by \code{\link[=make_data]{make_data()}}}

\item{n}{Sample size of actual or test coverage data}

\item{d.lower}{A numeric value for the lower classification threshold}

\item{d.upper}{A numeric value for the upper classification threshold}
}
\value{
A list of coverage proportions and LQAS outcomes
}
\description{
Function to perform LQAS based on data based on specified decision rules
}
\examples{
run_lqas(data = make_data(proportion = 0.3, pop = 10000),
         n = 40, d.lower = 60, d.upper = 90)

}
