% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-test_classifier.R
\name{simul_lqas}
\alias{simul_lqas}
\title{Function to perform a series of LQAS analysis on simulated coverage survey
data}
\usage{
simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)
}
\arguments{
\item{runs}{Number of simulation runs to perform per coverage proportion.
Default is 50 runs}

\item{pop}{Population size from which simulated coverage survey data is to
be taken from}

\item{n}{Sample size of actual or test coverage data}

\item{d.lower}{A numeric value for the lower classification threshold}

\item{d.upper}{A numeric value for the upper classification threshold}

\item{p.lower}{Starting proportion for simulations. Default is 0}

\item{p.upper}{Ending proportion for simulations. Default is 100}

\item{fine}{Granularity of simulated proportions; Default to 1}

\item{progress}{Logical. Should simulation progress be shown?
Default is TRUE}
}
\value{
A data.frame of coverage proportions and LQAS outcomes
}
\description{
Function to perform a series of LQAS analysis on simulated coverage survey
data
}
\examples{
simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)

}
