% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleacr.R
\docType{package}
\name{sleacr}
\alias{sleacr}
\alias{sleacr-package}
\title{sleacr: Simplified Lot Quality Assurance Sampling Evaluation of Access and Coverage (SLEAC) Tools in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

In the recent past, measurement of coverage has been mainly through two-stage cluster sampled surveys either as part of a nutrition assessment or through a specific coverage survey known as Centric Systematic Area Sampling (CSAS). However, such methods are resource intensive and often only used for final programme evaluation meaning results arrive too late for programme adaptation. SLEAC, which stands for Simplified Lot Quality Assurance Sampling Evaluation of Access and Coverage, is a low resource method designed specifically to address this limitation and is used regularly for monitoring, planning and importantly, timely improvement to programme quality, both for agency and Ministry of Health (MoH) led programmes. SLEAC is designed to complement the Semi-quantitative Evaluation of Access and Coverage (SQUEAC) method. This package provides functions for use in conducting a SLEAC assessment.
}
\references{
Myatt, Mark et al. 2012. Semi-Quantitative Evaluation of Access and
Coverage (SQUEAC)/Simplified Lot Quality Assurance Sampling Evaluation of
Access and Coverage (SLEAC) Technical Reference. Washington, DC: FHI 360/FANTA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutriverse.io/sleacr/}
  \item \url{https://github.com/nutriverse/sleacr}
  \item Report bugs at \url{https://github.com/nutriverse/sleacr/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

}
\keyword{internal}
