% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{SLEAC survey data from Sierra Leone}
\format{
A tibble with 14 rows and 6 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{country} \tab Country \cr
   \emph{province} \tab Province \cr
   \emph{district} \tab District \cr
   \emph{cases_in} \tab SAM cases found who are in the programme \cr
   \emph{cases_out} \tab SAM cases found who are not in the programme \cr
   \emph{rec_in} \tab Recovering SAM cases in the programme \cr
   \emph{cases_total} \tab Total number SAM cases found \cr
}
}
\source{
Ministry of Health, Sierra Leone
}
\usage{
survey_data
}
\description{
SLEAC survey data from Sierra Leone
}
\examples{
survey_data

}
\keyword{datasets}
