% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-test_classifier.R
\name{test_lqas_classifier}
\alias{test_lqas_classifier}
\title{Function to test performance of LQAS classifier}
\usage{
test_lqas_classifier(replicates = 20, runs = 50, pop = NULL,
  n = NULL, d.lower = NULL, d.upper = NULL, p.lower = 0,
  p.upper = 100, fine = 1, progress = TRUE)
}
\arguments{
\item{replicates}{Number of replicate LQAS simulations to perform.
Default is set to 20 replicates}

\item{runs}{Number of simulation runs to perform per coverage proportion.
Default is 50 runs}

\item{pop}{Population size from which simulated coverage survey data is to
be taken from}

\item{n}{Sample size of actual or test coverage data}

\item{d.lower}{A numeric value for the lower classification threshold}

\item{d.upper}{A numeric value for the upper classification threshold}

\item{p.lower}{Starting proportion for simulations. Default is 0}

\item{p.upper}{Ending proportion for simulations. Default is 100}

\item{fine}{Granularity of simulated proportiongs; Defaul to 1}

\item{progress}{Logical. Should simulation progress be shown? Default is TRUE}
}
\value{
A sleac object
}
\description{
Function to test performance of LQAS classifier
}
\examples{

test_lqas_classifier(pop = 10000, n = 40, d.lower = 60, d.upper = 90)

}
