% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{village_list}
\alias{village_list}
\title{List of villages in Bo District, Sierra Leone}
\format{
A tibble with 1001 rows and 4 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{id} \tab Unique identifier \cr
   \emph{chiefdom} \tab Chiefdom \cr
   \emph{section} \tab Section \cr
   \emph{village} \tab Village \cr
}
}
\source{
Ministry of Health, Sierra Leone
}
\usage{
village_list
}
\description{
List of villages in Bo District, Sierra Leone
}
\examples{
village_list

}
\keyword{datasets}
