% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-sample_size.R
\name{get_binom_hypergeom}
\alias{get_binom_hypergeom}
\title{Calculate the binomial coefficient "n-choose-k"}
\usage{
get_binom_hypergeom(n, k)
}
\arguments{
\item{n}{Total population}

\item{k}{Number of sample drawn from total population}
}
\value{
A numeric vector of binomial probability
}
\description{
Calculate the binomial coefficient "n-choose-k"
}
\examples{
get_binom_hypergeom(n = 600, k = 40)

}
