% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-sample_size.R
\name{get_hypergeom}
\alias{get_hypergeom}
\title{Calculate hypergeometric probability}
\usage{
get_hypergeom(k, m, n, N)
}
\arguments{
\item{k}{Number of cases in the sample}

\item{m}{Number of cases in the population}

\item{n}{Sample size}

\item{N}{Population size}
}
\value{
A numeric value of hypergeometric probability given specified
  parameters
}
\description{
Calculate hypergeometric probability
}
\examples{
get_hypergeom(k = 5, m = 600, n = 25, N = 10000)

}
