% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-classify_coverage.R
\name{lqas_classify_}
\alias{lqas_classify_}
\alias{lqas_classify}
\alias{lqas_classify_coverage}
\title{LQAS classifier}
\usage{
lqas_classify_(n, n_total, threshold = c(0.2, 0.5))

lqas_classify(n, n_total, threshold = c(0.2, 0.5))

lqas_classify_coverage(n, n_total, threshold = c(0.2, 0.5))
}
\arguments{
\item{n}{Number of cases found.}

\item{n_total}{Number sampled.}

\item{threshold}{Decision rule threshold/s. Should be between 0 and 1. At
least one threshold should be provided for a two-tier classifier. Two
thresholds should be provided for a three-tier classifier. Default is a
three-tier classifier with rule set at 0.2 and 0.5.}
}
\value{
A character value or vector indicating classification. If
\code{threshold} is a single value, the generic function returns \emph{1} if \code{n} is
greater than the threshold else \emph{0}. The coverage classifier
function returns \strong{"Satisfactory"} if \code{n} is greater than the threshold
else \strong{"Not satisfactory"}. If \code{threshold} is two values, the generic
function returns \emph{1} if \code{n} is greater than the first threshold and \emph{2} if
\code{n} is greater than the second threshold else \emph{0}. The CMAM coverage
classifier returns \strong{"Low"} if \code{n} is below or equal to lower threshold,
\strong{"High"} if \code{n} is above the higher threshold, and \strong{"Moderate"} for
all other values of \code{n}.
}
\description{
LQAS classifier
}
\examples{
lqas_classify_coverage(n = 6, n_total = 40)
with(
  survey_data,
  lqas_classify_coverage(n = cases_in, n_total = cases_total)
)

}
\author{
Ernest Guevarra
}
