% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-test_classifier.R
\name{plot.lqasSim}
\alias{plot.lqasSim}
\title{\code{plot} helper function for \code{test_lqas_classifier} function}
\usage{
\method{plot}{lqasSim}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lqasSim} produced by
\code{test_lqas_classifier} function}

\item{...}{Additional \code{plot()} parameters}
}
\value{
An LQAS probability of classification plot
}
\description{
\code{plot} helper function for \code{test_lqas_classifier} function
}
\examples{
x <- test_lqas_classifier(replicates = 5, runs = 5,
                          pop = 10000, n = 40, d.lower = 60, d.upper = 90)
plot(x)

}
