% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{SLEAC survey data from Sierra Leone}
\format{
A tibble with 14 rows and 6 columns:
\describe{
\item{\code{country}}{Country}
\item{\code{province}}{Province}
\item{\code{district}}{District}
\item{\code{in_cases}}{Cases found who are in the programme}
\item{\code{out_cases}}{Cases found who are not in the programme}
\item{\code{n}}{Total number of under 5 children sampled}
}
}
\source{
Ministry of Health, Sierra Leone
}
\usage{
survey_data
}
\description{
SLEAC survey data from Sierra Leone
}
\examples{
survey_data

}
\keyword{datasets}
