% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{village_list}
\alias{village_list}
\title{List of villages in Bo District, Sierra Leone}
\format{
A tibble with 1001 rows and 4 columns:
\describe{
\item{\code{id}}{Unique identifier}
\item{\code{chiefdom}}{Chiefdom}
\item{\code{section}}{Section}
\item{\code{village}}{Village}
}
}
\source{
Ministry of Health, Sierra Leone
}
\usage{
village_list
}
\description{
List of villages in Bo District, Sierra Leone
}
\examples{
village_list

}
\keyword{datasets}
